Fix possible AD NaN
Possible AD NaN would show up if f0==f1 which would be the case if the derivative wasn't dependent on state, and the initial dt routine was used, and if time was a dual number. While extremely rare, this could show up in cases where a callback is used to initiate the problem later in the integration.