Double the small ODE solver speed and remove usage of libm
For small enough ODEs, the `^` operation in the PI-controller actually dominates the runtime. You can see it in the profiles. This is why that was done with `@fastmath` which is the only portion of the ODE solver that does that, and thus reduces reproducibility. This replaces the fast `^` with an actual `^` approximation, and the effect is essentially a 2x speedup on small equations:

```julia
using OrdinaryDiffEq, BenchmarkTools
function lorenz(du,u,p,t)
 du[1] = 10.0(u[2]-u[1])
 du[2] = u[1]*(28.0-u[3]) - u[2]
 du[3] = u[1]*u[2] - (8/3)*u[3]
end
u0 = [1.0;0.0;0.0]
tspan = (0.0,100.0)
prob = ODEProblem(lorenz,u0,tspan)
@btime sol = solve(prob,Tsit5())

#=
Fast:
Length: 1318
400.499 μs (66 allocations: 6.27 KiB)

^:
Length: 1287
729.701 μs (66 allocations: 6.27 KiB)
=#

using Profile
Profile.clear()
@profile for i in 1:100 sol = solve(prob,Tsit5(),save_everystep=false) end
Juno.profiler()
```

The question I have is whether it's accurate enough to be relied on (https://discourse.julialang.org/t/slow-arbitrary-base-exponentiation-a-b/25386/20) but since it's just for an error estimate calculation which already is very hand-wavy it really don't need to be that accurate (just "good enough")