Fix mass matrix initialization algorithms for DAE vars not with eqs
In all of our previous test cases the equation that's supposed to define the algebraic variable matched with the algebraic variable, but that isn't necessarily the case. when looking at the residuals we need to look at the algebraic equations, and when looking at the variables to change we need to use the algebraic variables.