Don't differentiate the step controller
Odd little issue. The PI-controller in the stepper is not actually differentiable, so we need to keep the ties in the differentiation in the dts and t, but when we hit the qold gain on the integral control, we want to drop the AD there since we want to just have the chosen dt there. This fixes a fairly obscured bug that was found in one case of Pumas

>okay, tlag is fixed, and I also fixed the nan-safeness I think, and that also could fix the estimation robustness. Essentially, when differentiating something that also involves differentiating a callback, if a step rejected twice in a row where the error amount was less than qmax, and some of the states being differentiated have a zero derivative w.r.t. one of the parameters. I.e., this is the edge case-iest of edge cases, so it's hard to tell if you ever hit it in the optimization, but now you won't.