Fix ElasticArray tests
This is kind of a hack, but it looks like there was a major regression in that library and I'm not quite sure why, but given I found this while working on https://github.com/SciML/SciMLBase.jl/pull/1 I'm putting a fix to ElasticArrays right inline and will upstream that later. I'm kind of confused because I don't see a recent tag from ElasticArrays but it could've been a versioning thing.