Enforce correct saveat after tstop change
This is an odd one. If the user has the final time point in the `saveat` array, then `save_end == true`. Then if you `add_tstop!`, the final time point is no longer the final `saveat` which broke an assumption in the system. This decouples the `save_end` handling from the `saveat` mechanism so that if this `tstop` changes, it doesn't save the wrong point. It then needs to store whether the user overrode `save_end` (vs `save_end` as calculated by the default) in order to allow saving that new and improved endpoint.

A similar change is needed in StochasticDiffEq.jl, DelayDiffEq.jl, and StochasticDelayDiffEq.jl since they are vulnerable to the same issue.