Bypass autoswitch allocs on fully FSAL composites
From profiling I noticed that this was the main hotspot for composite algorithm overhead on small problems. `all_fsal` specializes on the algorithms and will hardcode a `true` if all of the algorithms are FSAL. This is true for common combinations like `AutoTsit5(Rosenbrock23())` and `AutoVern7(Rodas5())`, which then bypasses the dynamic dispatch (on the other algorithms, it just compiles to false in which case it's no overhead).

On a Lorenz system this is justified:

```julia
using OrdinaryDiffEq, StaticArrays
function lorenz(u,p,t)
  SA[10.0(u[2]-u[1]),u[1]*(28.0-u[3]) - u[2],u[1]*u[2] - (8/3)*u[3]]
end
u0 = SA[1.0;0.0;0.0]
tspan = (0.0,100.0)
prob = ODEProblem(lorenz,u0,tspan)
sol = solve(prob,Tsit5())
using Plots; plot(sol,vars=(1,2,3))

using BenchmarkTools
@btime sol = solve(prob,Tsit5(),save_everystep=false)
@btime sol = solve(prob,AutoTsit5(Rosenbrock23()),save_everystep=false)
```

```
Before: 284.100 μs (2609 allocations: 85.19 KiB)
After: 248.500 μs (57 allocations: 5.44 KiB)
No AutoSwitch: 215.700 μs (21 allocations: 3.67 KiB)
```

The remaining allocations seem to come from the tableau cache, which of course is hard to get rid of (and cannot necessarily be completely removed). There doesn't seem to be one source, and instead it seems to just be a small overhead which goes away for larger `f`