Better NaN detection
Fixes a hard to isolate case from https://discourse.julialang.org/t/incorporating-forcing-functions-in-the-ode-model/70133/26 . In there, ForwardDiff Duals were used, and the values were fine while the partials were NaN. This resulted in failed interpolations, which should instead have been caught by the isnan checks. isnan only checks the values, so that is the reason for the issue. By moving the isnan check to be on the normed value, this means that if what the user cares about isnan, then it will appropriately exit immediately.