Use a horrible extract hack to minimize footprint
structs are pass-by-value. While large structs are often passed by reference as
an optimization, in many cases they were passed by value instead,
causing a great deal of code to be emitted in order to copy the memory
around on the stack.
By inlining the caches to where they are used instead, we eliminate
the copies.

To minimize the amount of code we need to actually change, this commit
implements a macro that makes use of the old code with minimal
optimizations in order to define the associated variables. This
perhaps isn't the solution we want long-term, but it has the benefit
of allowing us to still take advantage of multiple dispatch in
defining the constants, so we can (for example) have `Float64` vs
`BigFloat` specializations.