Set u_modified in DAE initialization if the u was changed
Some integrators (such as Sundials) keep a shadow copy of `u` and need
to reflect this copy back into the solver internals to use them.
This happens e.g. during callbacks, but the same situation applies
to initialization, so use the same mechanism (setting u_modified
to `true`) to let the solver know that it needs to update its
internal datastructures.