Remove early defaulting and fix factorization algs
No longer default to GMRES on SplitODEProblem when it's not an operator equation. LUFactorization works fine. GenericFactorization has stronger assumptions than is required.

This removes the early defaulting since `linsolve=nothing` is type-inferrable for defaultalg since it's a single algorithm, and thus doing it early has no benefit but significant drawbacks. One drawback case is Radau since it picks the default based on real numbers but requires complex numbers. This is also the reason for the aforementioned factorization issue on SplitODEProblem, it was simply choosing wrong.