Update NonlinearSolve ForwardDiff support for DAE Initialization diff
NonlinearSolve.jl no longer differentiates the solver when used with ForwardDiff and instead solves and applies the chain rule at the solved point. This of course is a good thing, it's much more efficient. However, the DAE initialization algorithms were setup assuming that the element type of the nonlinear solve matches that of the ODE solver, which is violated by this change. This PR thus changes the caches to be non-dual'd to match what NonlinearSolve is expecting, and improves the testing of AD on initialization and solving of DAEs