Better warning on lack of initialization algorithm and slim Rosenbrock
This fixes the overly large Rosenbrock by giving and erroring mechanism to tell users that adding a nonlinear solver is required for DAEs. Before, we simply put it as a dep to Rosenbrock to fix the issues with the split PR with a note that this would be removed shortly. This does the removal and instead throws a contextual error saying you just need to add that library if you want that functionality. This should make Rosenbrocks a lot leaner for ODEs, and this will make cases where DAE intiialization is required on an explicit method give a much better error. In some cases, an error will go away since the trivial case is now contained in Core, meaning that if a nonlinear solver isn't actually required (analytically solved initialization) it will just handle it without erroring whereas before it would require the nonlinear solver libraries (and then not use them).