Fix some tests
The justification for this flip is rather deep... it's documented in https://github.com/EnzymeAD/Enzyme.jl/issues/2360 . Basically `using Enzyme` first causes some behavioral changes in the nonlinear solver. The fact that it matters is almost random here, so I am just isolating the solver tests from this for now.