Bump SciMLOperators
The way OrdinaryDiffEq.jl uses SciMLOperators is unchanged by the v1, which is on purpose because we knew we wanted to do this interface change for years. While this isn't the end of the story, we should make WOperator an AbstractSciMLOperator to simplify code, this is at least a valid bump so we should do it now.