New method Rodas6P
The new Rodas6P method was designed similarly to Rodas5P. It has order 6 for ODEs and index-1 DAEs with integrated order 5 interpolation.
232 order conditions must be satisfied, plus 175 for the interpolation.
Therefore, the number of stages is s=19, with the last 3 stages only required for the interpolation.
For large systems of stiff ODEs or DAEs with high accuracy requirements, the efficiency may be better than Rodas5P.